+++ R/difNLR-package.R
@@ -45,9 +45,9 @@
 #' Package: difNLR\cr
 #' Type: Package\cr
 #' Version: 1.4.2\cr
-#' Date: 2023-03-20\cr
+#' Date: 2023-03-22\cr
 #' Depends: R (>= 3.1)\cr
-#' Imports: calculus, ggplot2 (>= 2.2.1), msm, nnet, plyr, stats, VGAM\cr
+#' Imports: calculus, ggplot2 (>= 3.4.0), msm, nnet, plyr, stats, VGAM\cr
 #' Suggests: ShinyItemAnalysis\cr
 #' License: GPL-3\cr
 #' BugReports: \url{https://github.com/adelahladka/difNLR/issues}\cr